+++ tests/test_draw.py
@@ -1,30 +1,29 @@
 # vim: set fileencoding=utf-8 :
-import unittest
+import pytest
 
 import pyvips
-from .helpers import PyvipsTester
 
 
-class TestDraw(PyvipsTester):
+class TestDraw:
     def test_draw_circle(self):
         im = pyvips.Image.black(100, 100)
         im = im.draw_circle(100, 50, 50, 25)
         pixel = im(25, 50)
-        self.assertEqual(len(pixel), 1)
-        self.assertEqual(pixel[0], 100)
+        assert len(pixel) == 1
+        assert pixel[0] == 100
         pixel = im(26, 50)
-        self.assertEqual(len(pixel), 1)
-        self.assertEqual(pixel[0], 0)
+        assert len(pixel) == 1
+        assert pixel[0] == 0
 
         im = pyvips.Image.black(100, 100)
         im = im.draw_circle(100, 50, 50, 25, fill=True)
         pixel = im(25, 50)
-        self.assertEqual(len(pixel), 1)
-        self.assertEqual(pixel[0], 100)
+        assert len(pixel) == 1
+        assert pixel[0] == 100
         pixel = im(26, 50)
-        self.assertEqual(pixel[0], 100)
+        assert pixel[0] == 100
         pixel = im(24, 50)
-        self.assertEqual(pixel[0], 0)
+        assert pixel[0] == 0
 
     def test_draw_flood(self):
         im = pyvips.Image.black(100, 100)
@@ -32,10 +31,10 @@ class TestDraw(PyvipsTester):
         im = im.draw_flood(100, 50, 50)
 
         im2 = pyvips.Image.black(100, 100)
-        im2 = im.draw_circle(100, 50, 50, 25, fill=True)
+        im2 = im2.draw_circle(100, 50, 50, 25, fill=True)
 
         diff = (im - im2).abs().max()
-        self.assertEqual(diff, 0)
+        assert diff == 0
 
     def test_draw_image(self):
         im = pyvips.Image.black(51, 51)
@@ -48,17 +47,17 @@ class TestDraw(PyvipsTester):
         im3 = im3.draw_circle(100, 50, 50, 25, fill=True)
 
         diff = (im2 - im3).abs().max()
-        self.assertEqual(diff, 0)
+        assert diff == 0
 
     def test_draw_line(self):
         im = pyvips.Image.black(100, 100)
         im = im.draw_line(100, 0, 0, 100, 0)
         pixel = im(0, 0)
-        self.assertEqual(len(pixel), 1)
-        self.assertEqual(pixel[0], 100)
+        assert len(pixel) == 1
+        assert pixel[0] == 100
         pixel = im(0, 1)
-        self.assertEqual(len(pixel), 1)
-        self.assertEqual(pixel[0], 0)
+        assert len(pixel) == 1
+        assert pixel[0] == 0
 
     def test_draw_mask(self):
         mask = pyvips.Image.black(51, 51)
@@ -71,7 +70,7 @@ class TestDraw(PyvipsTester):
         im2 = im2.draw_circle(100, 50, 50, 25, fill=True)
 
         diff = (im - im2).abs().max()
-        self.assertEqual(diff, 0)
+        assert diff == 0
 
     def test_draw_rect(self):
         im = pyvips.Image.black(100, 100)
@@ -82,7 +81,7 @@ class TestDraw(PyvipsTester):
             im2 = im2.draw_line(100, 25, y, 74, y)
 
         diff = (im - im2).abs().max()
-        self.assertEqual(diff, 0)
+        assert diff == 0
 
     def test_draw_smudge(self):
         im = pyvips.Image.black(100, 100)
@@ -95,8 +94,8 @@ class TestDraw(PyvipsTester):
         im4 = im2.draw_image(im3, 10, 10)
 
         diff = (im4 - im).abs().max()
-        self.assertEqual(diff, 0)
+        assert diff == 0
 
 
 if __name__ == '__main__':
-    unittest.main()
+    pytest.main()