+++ tests/test_iofuncs.py
@@ -1,11 +1,11 @@
 # vim: set fileencoding=utf-8 :
-import unittest
+import pytest
 
 import pyvips
-from .helpers import PyvipsTester
+from helpers import assert_equal_objects
 
 
-class TestIofuncs(PyvipsTester):
+class TestIofuncs:
     # test the vips7 filename splitter ... this is very fragile and annoying
     # code with lots of cases
     def test_split7(self):
@@ -34,7 +34,7 @@ class TestIofuncs(PyvipsTester):
         ]
 
         for case in cases:
-            self.assertEqualObjects(split(case[0]), case[1])
+            assert_equal_objects(split(case[0]), case[1])
 
     def test_new_from_image(self):
         im = pyvips.Image.mask_ideal(100, 100, 0.5,
@@ -42,51 +42,52 @@ class TestIofuncs(PyvipsTester):
 
         im2 = im.new_from_image(12)
 
-        self.assertEqual(im2.width, im.width)
-        self.assertEqual(im2.height, im.height)
-        self.assertEqual(im2.interpretation, im.interpretation)
-        self.assertEqual(im2.format, im.format)
-        self.assertEqual(im2.xres, im.xres)
-        self.assertEqual(im2.yres, im.yres)
-        self.assertEqual(im2.xoffset, im.xoffset)
-        self.assertEqual(im2.yoffset, im.yoffset)
-        self.assertEqual(im2.bands, 1)
-        self.assertEqual(im2.avg(), 12)
+        assert im2.width == im.width
+        assert im2.height == im.height
+        assert im2.interpretation == im.interpretation
+        assert im2.format == im.format
+        assert im2.xres == im.xres
+        assert im2.yres == im.yres
+        assert im2.xoffset == im.xoffset
+        assert im2.yoffset == im.yoffset
+        assert im2.bands == 1
+        assert im2.avg() == 12
 
         im2 = im.new_from_image([1, 2, 3])
 
-        self.assertEqual(im2.bands, 3)
-        self.assertEqual(im2.avg(), 2)
+        assert im2.bands == 3
+        assert im2.avg() == 2
 
     def test_new_from_memory(self):
         s = bytearray(200)
         im = pyvips.Image.new_from_memory(s, 20, 10, 1, 'uchar')
 
-        self.assertEqual(im.width, 20)
-        self.assertEqual(im.height, 10)
-        self.assertEqual(im.format, 'uchar')
-        self.assertEqual(im.bands, 1)
-        self.assertEqual(im.avg(), 0)
+        assert im.width == 20
+        assert im.height == 10
+        assert im.format == 'uchar'
+        assert im.bands == 1
+        assert im.avg() == 0
 
         im += 10
 
-        self.assertEqual(im.avg(), 10)
+        assert im.avg() == 10
 
+    @pytest.mark.skipif(not pyvips.at_least_libvips(8, 5),
+                        reason="requires libvips >= 8.5")
     def test_get_fields(self):
-        if pyvips.at_least_libvips(8, 5):
-            im = pyvips.Image.black(10, 10)
-            fields = im.get_fields()
-            # we might add more fields later
-            self.assertTrue(len(fields) > 10)
-            self.assertEqual(fields[0], 'width')
+        im = pyvips.Image.black(10, 10)
+        fields = im.get_fields()
+        # we might add more fields later
+        assert len(fields) > 10
+        assert fields[0] == 'width'
 
     def test_write_to_memory(self):
         s = bytearray(200)
         im = pyvips.Image.new_from_memory(s, 20, 10, 1, 'uchar')
         t = im.write_to_memory()
 
-        self.assertEqual(s, t)
+        assert s == t
 
 
 if __name__ == '__main__':
-    unittest.main()
+    pytest.main()