+++ CHANGELOG.rst
@@ -7,6 +7,7 @@
 * add optional output params to docs [kleis]
 * update docs
 * add some libvips 8.7 tests
+* move to pytest [kleis]
 
 ## Version 2.1.2 (1 March 2018)